service/telegram-bot/requirements.txt
python-telegram-bot==20.6
django-environ==0.8.1
celery==5.3.4
loguru==0.6.0
redis==5.0.0
requests==2.26.0